requirements.txt
pywinauto
pygetwindow
pyyaml